+++ api/CHANGELOG.md
@@ -18,6 +18,7 @@
 - Pressing the system back buton on the `Apps` tab now exits the app
 - Made search queries and filters on the custom file picker persist across sessions https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/209
 - Improved export filenames for patch selections and keystores https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/205
+- Added persistent bundle sorting to the `Patch Bundle Discovery` https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/206
 
 
 # Bug fixes